check if file exist before removal
this prevents error message caused by trying to remove non-existent file

+++ rebuild.sh
@@ -11,5 +11,8 @@ hms=`date +%Y.%m.%d-%H%M`
 branch=`(git branch 2>/dev/null | sed -n '/^\* / { s|^\* ||; p; }')`
 revision=`(git rev-parse HEAD)`
 nixversion=`nixos-version | sed 's/\(.*\) .*/\1_/'`
-rm ~/.gtkrc-2.0
+
+if [ -f '~/.gtkrc-2.0' ]; then
+  rm ~/.gtkrc-2.0
+fi
 sudo NIXOS_LABEL="$nixversion$hms.$branch-${revision:0:7}" nixos-rebuild switch --impure --flake path:$SCRIPT_DIR#$system
